requirements.txt
alembic==1.5.8
apiron==5.1.0
Babel==2.9.0
barcode==1.0.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
beren==0.7.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
config==0.5.0.post0
cPython==0.0.6
crypto==1.4.1
cryptography==3.4.7
cycler
dataclasses
demjson3==3.0.6
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Cors==3.0.10
Flask-Ext==0.1
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
flask-tryton==0.10.0
Flask-WTF==0.14.3
forex-python==1.8
future==0.18.2
Genshi==0.7.6
geomet==0.3.0
greenlet
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
importlib-metadata==3.7.3
itsdangerous==1.1.0
Jinja2==2.11.3
jtutils==0.0.8
kiwisolver==1.3.1
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.4
Naked==0.1.31
nose==1.3.7
numpy
pandas==1.1.5
passlib==1.7.4
pendulum==2.1.2
Pillow
polib==1.1.0
proteus==6.0.4
pycairo==1.20.1
pycountry==20.7.3
pycparser==2.20
pycrypto==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.6
python-dotenv==0.20.0
python-editor==1.0.4
python-magic==0.4.22
python-sql==1.3.0
python-stdnum==1.16
pytz==2021.1
pytzdata==2020.1
PyYAML
qrcode==6.1
relatorio==0.10.0
requests==2.25.1
selenium==3.141.0
shellescape==3.8.1
simpleeval
simplejson==3.17.2
six==1.15.0
soupsieve==2.3.2.post1
SQLAlchemy==1.3.23
treepoem
tryton==6.0.13
typing-extensions==3.7.4.3
Unidecode==1.3.4
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
xlrd==2.0.1
xmltodict==0.13.0
zipp==3.4.1
